Backend/requirements.txt
#Webframework
fastapi[standard]==0.135.2
uvicorn[standard]==0.34.0
#DB
asyncpg==0.31.0
pydantic-settings==2.8.2
pydantic[email]==2.11.1
#Auth
argon2-cffi==25.1.0
PyJWT==2.10.1
httpx==0.28.1
#AI/RAG
sentence-transformers==5.1.1
chromadb==1.5.5
google-genai==1.68.0
#PDF
PyMuPDF==1.27.2.2
#Config
python-dotenv==1.1.0
#File-Handling
aiofiles==24.1.0
python-multipart==0.0.20